Refactoring HTTP client - added GetParams method.
